+++ src/ctests/profile_twoevents.c
@@ -36,7 +36,7 @@ int main(int argc, char **argv)
 
    if (start > end)
       test_fail(__FILE__, __LINE__, "Profile length < 0!", 0);
-   length = end - start;
+   length = (unsigned long)(end - start);
    prof_print_address("Test case profile: POSIX compatible profiling with two events.\n",prginfo);
    prof_print_prof_info(start,end,THRESHOLD,event_name);
    prof_alloc(2, length);
@@ -47,12 +47,12 @@ int main(int argc, char **argv)
    if (!TESTS_QUIET) {
       printf("Test type   : \tPAPI_PROFIL_POSIX\n");
    }
-   if ((retval = PAPI_profil(profbuf[0], blength, start, FULL_SCALE,
+   if ((retval = PAPI_profil(profbuf[0], (unsigned int)blength, start, FULL_SCALE,
                              EventSet, PAPI_event, THRESHOLD,
                              PAPI_PROFIL_POSIX)) != PAPI_OK) {
       test_fail(__FILE__, __LINE__, "PAPI_profil", retval);
    }
-   if ((retval = PAPI_profil(profbuf[1], blength, start, FULL_SCALE,
+   if ((retval = PAPI_profil(profbuf[1], (unsigned int)blength, start, FULL_SCALE,
                              EventSet, PAPI_TOT_CYC, THRESHOLD,
                              PAPI_PROFIL_POSIX)) != PAPI_OK)
       test_fail(__FILE__, __LINE__, "PAPI_profil", retval);
@@ -71,11 +71,11 @@ int main(int argc, char **argv)
       printf(TAB1, event_name, (values[1])[0]);
       printf(TAB1, "PAPI_TOT_CYC:", (values[1])[1]);
    }
-   if ((retval = PAPI_profil(profbuf[0], blength, start, FULL_SCALE,
+   if ((retval = PAPI_profil(profbuf[0], (unsigned int)blength, start, FULL_SCALE,
                              EventSet, PAPI_event, 0, PAPI_PROFIL_POSIX)) != PAPI_OK)
       test_fail(__FILE__, __LINE__, "PAPI_profil", retval);
 
-   if ((retval = PAPI_profil(profbuf[1], blength, start, FULL_SCALE,
+   if ((retval = PAPI_profil(profbuf[1], (unsigned int)blength, start, FULL_SCALE,
                              EventSet, PAPI_TOT_CYC, 0, PAPI_PROFIL_POSIX)) != PAPI_OK)
       test_fail(__FILE__, __LINE__, "PAPI_profil", retval);
